requirements.txt
# Core dependencies for Llama 3.2
llama-cpp-python>=0.2.0
transformers>=4.35.0
torch>=2.0.0

# For model loading and inference
huggingface-hub>=0.19.0
accelerate>=0.24.0

# CLI and interface
click>=8.1.0
prompt-toolkit>=3.0.0
rich>=13.0.0

# Configuration management
pyyaml>=6.0.0

# Stata syntax support
pygments>=2.16.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
